backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-cloud-storage>=2.18.0
google-cloud-aiplatform>=1.72.0
vertexai>=1.72.0
pydantic>=2.0.0
